🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(README.md): remove unnecessary content and update personal introduction
✨ feat(README.md): add section for showcasing skills
The unnecessary content in the README.md file has been removed to provide a cleaner and more concise introduction. The personal introduction has been updated to reflect the author's passion for automation and DevOps. A new section has been added to showcase the author's skills.

🐛 fix(ContactCTA.astro): update background color
The background color of the ContactCTA component has been updated to var(--gray-900) for better visual consistency.

🐛 fix(Icon.astro): update stop-color value
The stop-color value in the Icon.astro component has been updated to var(--accent-light) for better color consistency.

🐛 fix(IconPaths.ts): update github emoji name
The emoji name for the github logo in the IconPaths.ts file has been updated to "github" for better clarity and adherence to the GitMoji convention.

🐛 fix(Nav.astro): comment out ThemeToggle component
The ThemeToggle component in the Nav.astro file has been commented out to temporarily remove it from the navigation bar.

🐛 fix(Pill.astro): update background color
The background color of the Pill component has been updated to var(--gradient-accent-orange) for a more visually appealing design.

🐛 fix(Skills.astro): update background color
The background color of the Skills component has been updated to var(--gray-900) for better visual consistency.

✨ feat(career): add new career entry for Serviceware as Senior Software Engineer
A new career entry has been added for the position of Senior Software Engineer at Serviceware. The entry includes information about the company, the duration of the position, and the technologies used. The responsibilities of the role include overseeing the management, analysis, and maintenance of the portal application, as well as implementing strategies to enhance its functionality.

✨ feat