🎨 style(career/[...slug].astro): update max-width and styling for content section 🎨 style(global.css): update max-width for .wrapper class
The max-width for the .content class in career/[...slug].astro has been changed from 65ch to 90ch to allow for more content to be displayed within the specified width. Additionally, the font-size and margin for li elements within the .content class have been adjusted for better readability. The list-style property has been set to inside for ul elements within the .content class to improve the visual presentation of lists.

In global.css, the max-width for the .wrapper class has been changed from 83rem to 75rem to ensure consistent spacing and alignment across the application. The --link-color variable has been updated to use the --accent-regular color for better visual consistency.